+++ packages/ui/src/core/form-hint/form-hint.props.d.ts
@@ -2,5 +2,6 @@ import { InputSizeType } from '../input/input.props';
 
 export interface FormHintProps extends React.HTMLAttributes<HTMLSpanElement> {
   error?: boolean;
+  disabled?: boolean;
   size?: InputSizeType;
 }