+++ packages/ui/src/core/status-label/status-label.component.tsx
@@ -7,9 +7,12 @@ import { Size, Type, Variant } from './types';
 import Styled from './status-label.styles';
 
 const StatusLabel = intrinsicComponent<StatusLabelProps, HTMLDivElement>(
-  ({ label, variant = Variant.Filled, type = Type.Default, size, ...rest }: StatusLabelProps, ref): JSX.Element => (
+  (
+    { label, variant = Variant.Filled, type = Type.Default, size, iconSize = 8, ...rest }: StatusLabelProps,
+    ref
+  ): JSX.Element => (
     <Styled.StatusLabelWrapper ref={ref} size={size} label={label} type={type} variant={variant} {...rest}>
-      {type !== Type.Default && <Styled.StatusIcon type={type} />}
+      {type !== Type.Default && <Styled.StatusIcon type={type} iconSize={iconSize} />}
 
       <Styled.StatusLabel>{label}</Styled.StatusLabel>
     </Styled.StatusLabelWrapper>
@@ -27,6 +30,7 @@ StatusLabel.propTypes = {
   label: PT.node.isRequired,
   type: PT.oneOf(objectValues(Type)),
   size: PT.oneOf(objectValues(Size)),
+  iconSize: PT.number,
   variant: PT.oneOf(objectValues(Variant)),
 };